+++ src/banner/options.js
@@ -1,0 +1,15 @@
+class Options {
+    constructor(options) {
+        this.options = options;
+    }
+
+    has(optionName) {
+        return undefined !== this.options[optionName];
+    }
+
+    get(optionName, defaultValue = undefined) {
+        return this.options[optionName] || defaultValue;
+    }
+}
+
+module.exports = Options;