+++ src/template/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    single: require('./single'),
+    random: require('./random'),
+    multiple: require('./multiple'),
+};